Move interfaces from styles file to types file and add JSDocs
