+++ packages/office-ui-fabric-react/src/components/Coachmark/Coachmark.types.ts
@@ -1,4 +1,4 @@
-import { ICoachmarkStyles, ICoachmarkStyleProps } from './Coachmark.styles';
+import { IStyle } from '../../Styling';
 import { IPositioningContainerProps } from './PositioningContainer/PositioningContainer.types';
 import { IRefObject, IStyleFunctionOrObject } from '../../Utilities';
 import { Coachmark } from './Coachmark';
@@ -11,6 +11,7 @@ export interface ICoachmark {
   dismiss?: (ev?: any) => void;
 }
 
+/** Coachmark component props */
 export interface ICoachmarkProps extends React.ClassAttributes<Coachmark> {
   /**
    * Optional callback to access the ICoachmark interface. Use this instead of ref for accessing
@@ -168,6 +169,137 @@ export interface ICoachmarkProps extends React.ClassAttributes<Coachmark> {
    * Callback when the Coachmark tries to close.
    */
   onDismiss?: (ev?: any) => void;
+}
+
+/** The props needed to construct styles. */
+export interface ICoachmarkStyleProps {
+  /**
+   * Is the Coachmark collapsed.
+   * Deprecated, use `isCollapsed` instead.
+   * @deprecated Use `isCollapsed` instead.
+   */
+  collapsed?: boolean;
+
+  /**
+   * Is the Coachmark collapsed
+   */
+  isCollapsed: boolean;
+
+  /**
+   * Is the beacon currently animating.
+   */
+  isBeaconAnimating: boolean;
+
+  /**
+   * Is the component taking measurements
+   */
+  isMeasuring: boolean;
+
+  /**
+   * Is the Coachmark finished measuring the dimensions of innerHostElement
+   */
+  isMeasured: boolean;
+
+  /**
+   * The height measured before the component has been mounted
+   * in pixels
+   */
+  entityHostHeight?: string;
+
+  /**
+   * The width measured in pixels
+   */
+  entityHostWidth?: string;
+
+  /**
+   * Width of the coachmark
+   */
+  width?: string;
+
+  /**
+   * Height of the coachmark
+   */
+  height?: string;
+
+  /**
+   * Color
+   */
+  color?: string;
+
+  /**
+   * Beacon color one
+   */
+  beaconColorOne?: string;
+
+  /**
+   * Beacon color two
+   */
+  beaconColorTwo?: string;
+
+  /**
+   * Transform origin for teaching bubble content
+   */
+  transformOrigin?: string;
+
+  /**
+   * Delay time for the animation to start
+   */
+  delayBeforeCoachmarkAnimation?: string;
+}
+
+/** Represents the stylable areas of the control. */
+export interface ICoachmarkStyles {
+  /**
+   * Style for the root element in the default enabled/unchecked state.
+   */
+  root?: IStyle;
+
+  /**
+   * The pulsing beacon that animates when the Coachmark is collapsed.
+   */
+  pulsingBeacon?: IStyle;
+
+  /**
+   * The layer, or div, that the translate animation will be applied to.
+   */
+  translateAnimationContainer?: IStyle;
+
+  /**
+   * The layer the Scale animation will be applied to.
+   */
+  scaleAnimationLayer?: IStyle;
+
+  /**
+   * The layer the Rotate animation will be applied to.
+   */
+  rotateAnimationLayer?: IStyle;
+
+  /**
+   * The layer that content/components/elements will be hosted in.
+   */
+  entityHost?: IStyle;
+
+  /**
+   * The inner layer that components will be hosted in
+   * and primary purpose is scaling the layer down while the
+   * Coachmark collapsed.
+   */
+  entityInnerHost: IStyle;
+
+  /**
+   * The layer that directly contains the TeachingBubbleContent
+   */
+  childrenContainer: IStyle;
+
+  /**
+   * The styles applied when the Coachmark has collapsed.
+   */
+  collapsed?: IStyle;
+
+  /**
+   * The styles applied to the ARIA attribute container
+   */
+  ariaContainer?: IStyle;
 }
 
 /** @deprecated */
